meta-rcar-gen3: weston: Weston panel location and background color
This adds the following configuration variables that can be added
to local.conf to select Weston panel location and background color:
  WESTON_PANEL_LOCATION
  WESTON_BACKGROUND_COLOR

Panel location can be either "none", or "top" which is the default.
Background color should be in the "0xAARRGGBB" format. By default
a tiling background image is used instead of solid color.

Signed-off-by: Valentine Barshak <valentine.barshak@cogentembedded.com>