Fixed issues with IMX219 camera sensor
- Removed streaming control switching on/off sensor, CSI stops
recieving packets after a while if this is enabled. Still do not
like idea of clock always running, but found no better fix.
- Tuned analog controls of sensor (less dark and less white depending
on environment for gain and shutter)
- Set FIFO (shift register) divider value to proper value (8),
which fixed all artifacts with bayer output. This was burst of luck
to find this (there is no publically available information on how to
set this sensor in RAW8 mode except for this driver now).

Signed-off-by: Sergey Lapin <sergey.lapin@cogentembedded.com>