conf: layer.conf: Cleanup (move packages to packagegrups, etc).
This is based on the original package clean-up patch
by Andrey Dolnikov <andrey.dolnikov@cogentembedded.com>.
This moves packages to package groups. This also includes
some of the packages conditionally instead of making the
packages empty if the condition is not met.

While at it, add python2 package group to v3x, sort package
dependencies in the alphabetical order, and add the following
packages:

  packagegroup-bsp-devdbg:
    + dropbear (if IMAGE_FEATURES has no "ssh-server-openssh")
  packagegroup-bsp-devdbg:
    + atop
  packagegroup-bsp-utest:
    + linux-renesas-uapi
  packagegroup-v3x:
    + br-test

Signed-off-by: Valentine Barshak <valentine.barshak@cogentembedded.com>