recipes-kernel: linux-renesas: Allow forced DMA support
This patch set is needed for cmemdrv which uses of_dma_configure.
When a device is not located on either a platform or a PCI bus,
and of_dma_get_range() returns -ENODEV, dummy device operations
are set for the device. This patch drops platform and PCI bus
dependencies and allows to force DMA usage for a device.

Signed-off-by: Valentine Barshak <valentine.barshak@cogentembedded.com>