recipes-kernel: linux-renesas: Fix I2C DMA issues
Since we need to setup new messages in irq context,
we can't allocate bounce buffers if a message buffer
is not DMA-safe. Fall back to PIO in that case.

For example, this fixes infinite "SET TIME!" warnings
in the DS1307 RTC probe function which uses DMA-unsafe
transfer buffers allocated on the stack.

Signed-off-by: Valentine Barshak <valentine.barshak@cogentembedded.com>