recipes-core: systemd: Add a prefix to network configuration files
This adds a "70-" prefix to the network configuration files
to avoid them being overridden by the default configuration
files from /lib/systemd/network/ which have the "80-" prefix.

Otherwise, the 80-wired.network configuration is used for all
the wired connections in Yocto Project 2.8, and we need the
"nfsroot=" kernel command line option in order to keep the
NFS root working after starting the systemd-networkd daemon.

Signed-off-by: Valentine Barshak <valentine.barshak@cogentembedded.com>